renamed nineml.utility to nineml.utils to match other modules/packages
