+++ lib9ml/python/nineml/abstraction_layer/componentclass/utils/modifiers.py
@@ -8,7 +8,7 @@ This file contains utility classes for modifying components.
 from ..base import Parameter
 from .cloner import ComponentExpandPortDefinition
 from ...ports import AnalogSendPort, AnalogReducePort, AnalogReceivePort
-from nineml.utility import filter_expect_single
+from nineml.utils import filter_expect_single
 from nineml.exceptions import NineMLRuntimeError
