+++ lib9ml/python/nineml/abstraction_layer/dynamics/testing_utils/testable_component.py
@@ -9,8 +9,8 @@ import os
 import sys
 
 from nineml.exceptions import NineMLRuntimeError
-from nineml.utility import restore_sys_path
-from nineml.utility import LocationMgr
+from nineml.utils import restore_sys_path
+from nineml.utils import LocationMgr
 from nineml.abstraction_layer import DynamicsClass
