+++ lib9ml/python/nineml/abstraction_layer/dynamics/utils/equality_checker.py
@@ -5,7 +5,7 @@ docstring needed
 :license: BSD-3, see LICENSE for details.
 """
 
-from nineml.utility import safe_dict
+from nineml.utils import safe_dict
 from ...componentclass.utils.equality_checker import (
     ComponentEqualityChecker, assert_equal,
     assert_equal_list)
