+++ lib9ml/python/nineml/abstraction_layer/dynamics/validators/namingconflicts.py
@@ -4,7 +4,7 @@ docstring needed
 :copyright: Copyright 2010-2013 by the Python lib9ML team, see AUTHORS.
 :license: BSD-3, see LICENSE for details.
 """
-from nineml.utility import assert_no_duplicates
+from nineml.utils import assert_no_duplicates
 from nineml.exceptions import NineMLRuntimeError
 from ...componentclass.validators import (
     LocalNameConflictsComponentValidator,
