+++ lib9ml/python/nineml/abstraction_layer/ports.py
@@ -7,7 +7,7 @@ This file defines the Port classes used in NineML
 from abc import ABCMeta
 from . import BaseALObject
 from nineml.abstraction_layer.units import dimensionless
-from nineml.utility import ensure_valid_identifier
+from nineml.utils import ensure_valid_identifier
 from nineml.exceptions import NineMLRuntimeError
