+++ lib9ml/python/nineml/user_layer/values.py
@@ -2,7 +2,7 @@
 from . import BaseULObject
 from nineml.xmlns import E, NINEML
 from nineml.annotations import read_annotations, annotate_xml
-from ..utility import check_tag
+from nineml.utils import check_tag
 
 
 class BaseValue(BaseULObject):
